+++ gaphor/UML/actions/__init__.py
@@ -25,6 +25,7 @@ from gaphor.UML.actions.flow import ControlFlowItem, ObjectFlowItem
 from gaphor.UML.actions.objectnode import ObjectNodeItem
 from gaphor.UML.actions.partition import PartitionItem
 from gaphor.UML.actions.pin import InputPinItem, OutputPinItem
+import gaphor.UML.actions.drop
 
 
 __all__ = [